fragment capacity tests are fixed
